działa import

+++ src/main/java/com/janis/bgg/demo/xml/Items3/Result.java
@@ -1,20 +1,13 @@
-
 package com.janis.bgg.demo.xml.Items3;
 
-import java.math.BigInteger;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlSchemaType;
-import javax.xml.bind.annotation.XmlType;
+import javax.xml.bind.annotation.*;
 
 
 /**
  * <p>Java class for anonymous complex type.
- * 
+ *
  * <p>The following schema fragment specifies the expected content contained within this class.
- * 
+ *
  * <pre>
  * &lt;complexType>
  *   &lt;complexContent>
@@ -26,8 +19,6 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;/complexContent>
  * &lt;/complexType>
  * </pre>
- * 
- * 
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "")
@@ -35,68 +26,58 @@ import javax.xml.bind.annotation.XmlType;
 public class Result {
 
     @XmlAttribute(name = "level")
-    protected BigInteger level;
+    protected Integer level;
     @XmlAttribute(name = "numvotes", required = true)
-    protected BigInteger numvotes;
+    protected Integer numvotes;
     @XmlAttribute(name = "value", required = true)
     @XmlSchemaType(name = "anySimpleType")
     protected String value;
 
     /**
      * Gets the value of the level property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link BigInteger }
-     *     
+     *
+     * @return possible object is
+     * {@link Integer }
      */
-    public BigInteger getLevel() {
+    public Integer getLevel() {
         return level;
     }
 
     /**
      * Sets the value of the level property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link BigInteger }
-     *     
+     *
+     * @param value allowed object is
+     *              {@link Integer }
      */
-    public void setLevel(BigInteger value) {
+    public void setLevel(Integer value) {
         this.level = value;
     }
 
     /**
      * Gets the value of the numvotes property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link BigInteger }
-     *     
+     *
+     * @return possible object is
+     * {@link Integer }
      */
-    public BigInteger getNumvotes() {
+    public Integer getNumvotes() {
         return numvotes;
     }
 
     /**
      * Sets the value of the numvotes property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link BigInteger }
-     *     
+     *
+     * @param value allowed object is
+     *              {@link Integer }
      */
-    public void setNumvotes(BigInteger value) {
+    public void setNumvotes(Integer value) {
         this.numvotes = value;
     }
 
     /**
      * Gets the value of the value property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
+     *
+     * @return possible object is
+     * {@link String }
      */
     public String getValue() {
         return value;
@@ -104,11 +85,9 @@ public class Result {
 
     /**
      * Sets the value of the value property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
+     *
+     * @param value allowed object is
+     *              {@link String }
      */
     public void setValue(String value) {
         this.value = value;
